requirements.txt
keras==2.11.0
numpy==1.24.1
Pillow==9.4.0
streamlit==1.18.1
